I KNOW I KNOW I SHOULD BE MAKING COMPONENTS
I will refactor soon, but for now I am still trying to iron out proper transitioning. Also, I adapted the mobile and tablet views to fit better with the home screens.

Quite unfortunate that the vuetify transitions are quite bad, but I have managed to hack a nice solution which is pretty much seamless.

I'm still looking for some call to action to put on the slides, maybe a shortcut of sorts. It looks too sad with 2 lines of text on each slide. Gotta change that.